requirements.txt
Flask>=2.3
PyYAML>=6.0

streamlit>=1.30
